+++ CHANGES.md
@@ -1,4 +1,6 @@
 #3.3.0 beta4 2013-??-??
+ - 2013-09-24 Fixed bug#[9750](http://bugs.otrs.org/show_bug.cgi?id=9750) - Problem with SSO and security restricted iframes in IE8+WXP by adding SysConfig option *DisableMSIFrameSecurityRestricted*.
+    Thanks to Pawel @ ib.pl.
  - 2013-09-24 Added otrs.AddCustomerUser2Group.pl command line script.
  - 2013-09-20 Fixed bug#[9720](http://bugs.otrs.org/show_bug.cgi?id=9720) - CIC gives error message when clicking on phone ticket while that module is not registered.
  - 2013-09-20 Fixed bug#[9742](http://bugs.otrs.org/show_bug.cgi?id=9742) - Double close icon on TreeView for Customer Interface.